requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
python-dotenv==1.1.0
motor==3.7.0
pydantic==2.11.3
pymongo==4.12.0
openai